Constraints

The following are constraints that PowerModelsRestoration adds.

PowerModelsRestoration.restoration_comps — Constant

supported components for restoration

Base.isinteger — Method

All functions neccessary because z is sometimes an integer instead of variable_ref

JuMP.value — Method

Need because of mix of vars/real in zgen, zbranch, z_storage

PowerModels.constraint_power_magnitude_sqr_on_off — Method

p[arc_from]^2 + q[arc_from]^2 <= w[f_bus]/tm*ccm[i]

PowerModelsRestoration._get_damaged_items — Method

PowerModelsRestoration._get_isolated_load — Method

PowerModelsRestoration._get_repairable_items — Method

PowerModelsRestoration._propagate_damage_status! — Method

PowerModelsRestoration._ref_add_damaged_buses! — Method

PowerModelsRestoration._set_component_inactive! — Method

PowerModelsRestoration.build_mrsp — Method

PowerModelsRestoration.build_restoration_redispatch — Method

PowerModelsRestoration.build_rop — Method

PowerModelsRestoration.clean_solution! — Method

Replace NaN and Nothing with 0 in multinetwork solutions

PowerModelsRestoration.clean_status! — Method

Replace non-integer status codes for devices, maps bus status to bus_type

PowerModelsRestoration.clear_damage_indicator! — Method

Clear damage indicator and replace with status=0

PowerModelsRestoration.constraint_branch_bus_connection — Method

on/off constraint for branches connected to damaged buses

PowerModelsRestoration.constraint_branch_damage — Method

PowerModelsRestoration.constraint_branch_energized — Method

PowerModelsRestoration.constraint_bus_damage_soft — Method

PowerModelsRestoration.constraint_bus_damage_soft — Method

no vm values to turn off

PowerModelsRestoration.constraint_bus_damage_soft — Method

PowerModelsRestoration.constraint_bus_damage_soft — Method

PowerModelsRestoration.constraint_bus_energized — Method

PowerModelsRestoration.constraint_gen_bus_connection — Method

on/off constraint for generators connected to damaged buses

PowerModelsRestoration.constraint_gen_damage — Method

PowerModelsRestoration.constraint_gen_energized — Method

PowerModelsRestoration.constraint_load_bus_connection — Method

on/off constraint for loads connected to damaged buses

PowerModelsRestoration.constraint_load_damage — Method

PowerModelsRestoration.constraint_load_increasing — Method

Load delivered at each node must be greater than or equal the previous time period

PowerModelsRestoration.constraint_model_voltage_damage — Method

PowerModelsRestoration.constraint_model_voltage_damage — Method

PowerModelsRestoration.constraint_model_voltage_damage — Method

PowerModelsRestoration.constraint_ohms_yt_from_damage — Method

PowerModelsRestoration.constraint_ohms_yt_from_damage — Method

PowerModelsRestoration.constraint_ohms_yt_to_damage — Method

PowerModelsRestoration.constraint_ohms_yt_to_damage — Method

PowerModelsRestoration.constraint_power_balance_shed — Method

PowerModelsRestoration.constraint_power_balance_shed — Method

PowerModelsRestoration.constraint_power_balance_shed — Method

PowerModelsRestoration.constraint_power_balance_shed — Method

PowerModelsRestoration.constraint_restoration_cardinality_lb — Method

PowerModelsRestoration.constraint_restoration_cardinality_lb — Method

Limit the minimum number of items restored in each time-step

PowerModelsRestoration.constraint_restoration_cardinality_ub — Method

PowerModelsRestoration.constraint_restoration_cardinality_ub — Method

Limit the maximum number of items restored in each time-step

PowerModelsRestoration.constraint_shunt_bus_connection — Method

on/off constraint for shunts connected to damaged buses

PowerModelsRestoration.constraint_shunt_damage — Method

PowerModelsRestoration.constraint_storage_bus_connection — Method

on/off constraint for storage connected to damaged buses

PowerModelsRestoration.constraint_storage_damage — Method

PowerModelsRestoration.constraint_storage_energized — Method

PowerModelsRestoration.constraint_thermal_limit_from_damage — Method

constraint_thermal_limit_from(pm::AbstractPowerModel, n::Int, i::Int)

Adds the (upper and lower) thermal limit constraints for the desired branch to the PowerModel.

PowerModelsRestoration.constraint_thermal_limit_to_damage — Method

PowerModelsRestoration.constraint_voltage_angle_difference_damage — Method

PowerModelsRestoration.constraint_voltage_magnitude_bounds_soft — Method

PowerModelsRestoration.constraint_voltage_magnitude_bounds_soft — Method

no vm values to turn off

PowerModelsRestoration.constraint_voltage_magnitude_bounds_soft — Method

PowerModelsRestoration.constraint_voltage_magnitude_bounds_soft — Method

PowerModelsRestoration.constraint_voltage_magnitude_sqr_from_damage — Method

PowerModelsRestoration.constraint_voltage_magnitude_sqr_to_damage — Method

PowerModelsRestoration.constraint_voltage_product_damage — Method

PowerModelsRestoration.count_damaged_items — Method

Count the number of items with a key "damaged" == 1 in a network

PowerModelsRestoration.count_repairable_items — Method

PowerModelsRestoration.damage_items! — Method

Set damage status for damageditems in nwdata

PowerModelsRestoration.get_damaged_items — Method

PowerModelsRestoration.get_isolated_load — Method

PowerModelsRestoration.get_repairable_items — Method

PowerModelsRestoration.objective_min_restoration — Method

PowerModelsRestoration.print_summary_restoration — Method

prints a summary of a restoration solution to the terminal

PowerModelsRestoration.propagate_damage_status! — Method

propagates the damaged indicator from buses to indicent branches, generators, shunts, and storage

PowerModelsRestoration.ref_add_damaged_branches! — Method

PowerModelsRestoration.ref_add_damaged_gens! — Method

PowerModelsRestoration.ref_add_damaged_items! — Method

PowerModelsRestoration.ref_add_damaged_storage! — Method

PowerModelsRestoration.replicate_restoration_network — Method

Transforms a single network into a multinetwork with several deepcopies of the original network. Indexed from 0.

PowerModelsRestoration.replicate_restoration_network — Method

Transforms a single network into a multinetwork with several deepcopies of the original network. Indexed from 0.

PowerModelsRestoration.run_ac_mld_uc — Method

implements a scalable huristic solution to the AC-MLD problem

PowerModelsRestoration.run_mrsp — Method

PowerModelsRestoration.run_restoration_redispatch — Method

Simulate a restoration sequence power flow

PowerModelsRestoration.run_restoration_redispatch — Method

Simulate a restoration sequence power flow

PowerModelsRestoration.run_rop — Method

PowerModelsRestoration.set_component_inactive! — Method

PowerModelsRestoration.summary_restoration — Method

prints a summary of a restoration solution

PowerModelsRestoration.update_status! — Method

updates status in the data model with values from a sparse solution dict

PowerModelsRestoration.variable_branch_damage_indicator — Method

variable: 0 <= damage_branch[l] <= 1 for l in branches

PowerModelsRestoration.variable_bus_damage_indicator — Method

variable: 0 <= damage_bus[l] <= 1 for l in buses

PowerModelsRestoration.variable_bus_voltage_damage — Method

PowerModelsRestoration.variable_bus_voltage_damage — Method

PowerModelsRestoration.variable_bus_voltage_damage — Method

PowerModelsRestoration.variable_bus_voltage_magnitude_on_off — Method

variable: v[i] for i in buses

PowerModelsRestoration.variable_bus_voltage_magnitude_sqr_on_off — Method

PowerModelsRestoration.variable_bus_voltage_magnitude_sqr_violation — Method

PowerModelsRestoration.variable_bus_voltage_magnitude_violation — Method

variable: v[i] for i in buses

PowerModelsRestoration.variable_gen_damage_indicator — Method

variable: 0 <= damage_gen[l] <= 1 for l in genes

PowerModelsRestoration.variable_gen_power_damage — Method

generates variables for both active and reactive generation

PowerModelsRestoration.variable_gen_power_imaginary_damage — Method

apo models ignore reactive power flows

PowerModelsRestoration.variable_gen_power_imaginary_damage — Method

variable: qq[j] for j in gen

PowerModelsRestoration.variable_gen_power_real_damage — Method

variable: pg[j] for j in gen

PowerModelsRestoration.variable_storage_current_damage — Method

do nothing by default but some formulations require this

PowerModelsRestoration.variable_storage_current_damage — Method

this is the same as non-damaged version becouse ccms includes zero

PowerModelsRestoration.variable_storage_damage_indicator — Method

variable: 0 <= damage_storage[l] <= 1 for l in storagees

PowerModelsRestoration.variable_storage_power_control_imaginary_damage — Method

PowerModelsRestoration.variable_storage_power_imaginary_damage — Method

apo models ignore reactive power flows

PowerModelsRestoration.variable_storage_power_imaginary_damage — Method

PowerModelsRestoration.variable_storage_power_mi_damage — Method

PowerModelsRestoration.variable_storage_power_real_damage — Method