PowerModelsRestoration.run_restoration_redispatch — Function

Simulate a restoration sequence power flow

Simulate a restoration sequence power flow

Missing docstring.

Missing docstring for run_rop_heuristic. Check Documenter's build log for details.

PowerModelsRestoration._build_partial_rop — Method

PowerModelsRestoration._count_component_repairs — Method

count repairs in each period of the solution data

PowerModelsRestoration._count_cumulative_component_repairs — Method

count the cumulative repairs in each period of the solution data

PowerModelsRestoration._fill_missing_variables! — Method

    `_fill_missing_variables!(sol::Dict{String,<:Any}, network::Dict{String,<:Any})`
add network values for components removed for status=0 from solution

PowerModelsRestoration._run_partial_rop — Method

Variation of run_rop that permits some load not to be restored in final period. Useful for partial restoration of a network in rop heuristics

PowerModelsRestoration._single_repair_restoration_network — Function

Create a restoration network with 1 repair per time period. Final period will contain remainder repairs.

PowerModelsRestoration.get_component_activations — Method

get repairs in each period of the solution data

PowerModelsRestoration.merge_solution! — Method

Update sol1 dictionary with sol2 dictionary. termination_status, primal_status, and dual_status values are the maximum of the MOI status code. solvetime,objective, and objective_lb are accumulated. The solution dictionary is merged using PowerModels.update_data(sol_1["solution"],sol_2["solution"]).