Constraints

The following are constraints that PowerModelsRestoration adds.

PowerModelsRestoration.constraint_branch_damage — Method

PowerModelsRestoration.constraint_bus_damage_soft — Method

PowerModelsRestoration.constraint_gen_damage — Method

PowerModelsRestoration.constraint_load_damage — Method

PowerModelsRestoration.constraint_model_voltage_damage — Method

PowerModelsRestoration.constraint_ohms_yt_from_damage — Method

PowerModelsRestoration.constraint_ohms_yt_to_damage — Method

PowerModelsRestoration.constraint_power_balance_shed — Method

PowerModelsRestoration.constraint_restoration_cardinality_lb — Method

Limit the minimum number of components restored in each time-step

PowerModelsRestoration.constraint_restoration_cardinality_ub — Method

Limit the maximum number of components restored in each time-step

PowerModelsRestoration.constraint_shunt_damage — Method

PowerModelsRestoration.constraint_storage_damage — Method

PowerModelsRestoration.constraint_thermal_limit_from_damage — Method

constraint_thermal_limit_from(pm::AbstractPowerModel, n::Int, i::Int)

Adds the (upper and lower) thermal limit constraints for the desired branch to the PowerModel.

PowerModelsRestoration.constraint_thermal_limit_to_damage — Method

PowerModelsRestoration.constraint_voltage_angle_difference_damage — Method

PowerModelsRestoration.constraint_voltage_magnitude_bounds_soft — Method

PowerModelsRestoration.constraint_branch_bus_connection — Method

on/off constraint for branches connected to damaged buses

PowerModelsRestoration.constraint_branch_energized — Method

PowerModelsRestoration.constraint_bus_energized — Method

PowerModelsRestoration.constraint_gen_bus_connection — Method

on/off constraint for generators connected to damaged buses

PowerModelsRestoration.constraint_gen_energized — Method

PowerModelsRestoration.constraint_load_bus_connection — Method

on/off constraint for loads connected to damaged buses

PowerModelsRestoration.constraint_load_increasing — Method

Load delivered at each node must be greater than or equal the previous time period

PowerModelsRestoration.constraint_restoration_cardinality_lb — Method

PowerModelsRestoration.constraint_restoration_cardinality_ub — Method

PowerModelsRestoration.constraint_shunt_bus_connection — Method

on/off constraint for shunts connected to damaged buses

PowerModelsRestoration.constraint_storage_bus_connection — Method

on/off constraint for storage connected to damaged buses

PowerModelsRestoration.constraint_storage_energized — Method